feat: dexscreener = 0.1.2
