+++ solagent-plugins/dexscreener/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "solagent-plugin-dexscreener"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2021"
 authors = ["zTgx <beautifularea@gmail.com>"]
 repository = "https://github.com/zTgx/solagent.rs"
